+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 7.0.3
+
+## 7.0.2
+
+## 7.0.1
+
 ## 7.0.0
 
 ### Major Changes
@@ -12,7 +18,6 @@
 ### Patch Changes
 
 - Fix versioned icons to work with next versions. _[`#3186`](https://github.com/solid-design-system/solid/pull/3186) [`30af3ad`](https://github.com/solid-design-system/solid/commit/30af3ad6d2e8daebeaa9d6a863fdd5794a015da9) [@auroraVasconcelos](https://github.com/auroraVasconcelos)_
-
 
 ## 6.30.0